refactor(execution, native_blockifier): make TransactionExecutor.execute() work w/o Py objects
